refactor(examples): replace space-management with full-app-headless
