+++ examples/full-app-headless/index.html
@@ -3,11 +3,10 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Space Management Example - Storacha Console Toolkit</title>
+    <title>Storacha Console Toolkit - Full app (headless)</title>
   </head>
   <body>
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
-
