+++ examples/full-app-headless/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@storacha/console-toolkit-example-space-management",
+  "name": "@storacha/console-toolkit-example-full-app-headless",
   "private": true,
   "version": "0.1.0",
   "type": "module",
