+++ examples/full-app-headless/src/main.tsx
