+++ examples/full-app-headless/tsconfig.json
