+++ examples/full-app-headless/vite.config.ts
